+++ server/server/server.go
@@ -72,6 +72,7 @@ func StartServer(config *viper.Viper, flags map[string]interface{}) {
 	router.GET("/checkVersion", handler.FindLatestVersion)
 	router.GET("/apps/latest", handler.FetchLatestVersionOfApp)
 	router.GET("/download/latest/:owner/:app_identifier/:platform", handler.PublicLatestDownload)
+	router.GET("/d/:app/:platform", handler.ShortLatestDownload)
 	loginLimiter := utils.NewIPRateLimiter(rate.Every(6*time.Second), 10)
 	signupLimiter := utils.NewIPRateLimiter(rate.Every(20*time.Second), 3)
 	router.POST("/signup", utils.RateLimitMiddleware(signupLimiter), handler.SignUp)